# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: КадастрПро API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
...
  /orders/{orderId}/bids:
    get:
      operationId: listOrderBids
      tags: [bids]
      summary: Get bids for an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bid"
    post:
      operationId: createBid
      tags: [bids]
      summary: Submit bid on an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BidInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
  /orders/{orderId}/bids/{bidId}:
    patch:
      operationId: updateBid
      tags: [bids]
      summary: Accept or reject a bid
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
        - name: bidId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BidUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
  /engineers/{engineerId}/bids:
    get:
      operationId: listEngineerBids
      tags: [bids]
      summary: Get bids submitted by an engineer
      parameters:
        - name: engineerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bid"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        telegram:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isBlocked:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, email, role, createdAt]
...
    RegisterInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
        phone:
          type: string
        attestatNumber:
          type: string
      required: [name, email, password, role]
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResult:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    PortfolioItem:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
        region:
          type: string
        year:
          type: integer
        area:
          type: ["string", "null"]
      required: [title, description, type, region, year]
...
    Engineer:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        registryNumber:
          type: string
        specializations:
          type: array
          items:
            type: string
        region:
          type: string
        regions:
          type: array
          items:
            type: string
        district:
          type: ["string", "null"]
        sro:
          type: ["string", "null"]
        experience:
          type: integer
        bio:
          type: ["string", "null"]
        isVerified:
          type: boolean
        isOnline:
          type: boolean
        isHidden:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        completedOrders:
          type: integer
        responseTime:
          type: ["string", "null"]
        priceFrom:
          type: ["integer", "null"]
        portfolioItems:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioItem"
        isPro:
          type: boolean
        proExpiresAt:
          type: ["string", "null"]
        debtAmount:
          type: integer
        createdAt:
          type: string
        attestatNumber:
          type: ["string", "null"]
        rosreestrStatus:
          type: ["string", "null"]
        sroName:
          type: ["string", "null"]
        rosreestrCheckedAt:
          type: ["string", "null"]
        rosreestrWorksCount:
          type: ["integer", "null"]
        rosreestrRejectionsCount:
          type: ["integer", "null"]
        rosreestrSuspensionsCount:
          type: ["integer", "null"]
        rosreestrRejectionRate:
          type: ["number", "null"]
      required: [id, userId, user, registryNumber, specializations, region, regions, experience, isVerified, isOnline, isHidden, rating, reviewCount, completedOrders, portfolioItems, isPro, debtAmount, createdAt]
...
    Order:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customer:
          $ref: "#/components/schemas/User"
        title:
          type: string
        description:
          type: string
        serviceType:
          type: string
        region:
          type: string
        budget:
          type: ["number", "null"]
        deadline:
          type: ["string", "null"]
        status:
          type: string
        bidCount:
          type: integer
        createdAt:
          type: string
      required: [id, customerId, customer, title, description, serviceType, region, status, bidCount, createdAt]
...
    Bid:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        order:
          $ref: "#/components/schemas/Order"
        engineerId:
          type: integer
        engineer:
          $ref: "#/components/schemas/Engineer"
        message:
          type: string
        price:
          type: ["number", "null"]
        proposedDeadline:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
      required: [id, orderId, order, engineerId, engineer, message, status, createdAt]
    BidInput:
      type: object
      properties:
        message:
          type: string
        price:
          type: number
        proposedDeadline:
          type: string
      required: [message]
    BidUpdate:
      type: object
      properties:
        status:
          type: string
      required: [status]
...